feat: basic data model
