--- src/main/java/pl/put/poznan/sqc/logic/ActorAction.java
@@ -1,5 +1,0 @@
-package pl.put.poznan.sqc.logic;
-
-public class ActorAction {
-
-}
