--- src/main/java/pl/put/poznan/sqc/logic/CountKeywords.java
@@ -1,5 +1,0 @@
-package pl.put.poznan.sqc.logic;
-
-public class CountKeywords {
-
-}
