--- src/main/java/pl/put/poznan/sqc/logic/Download.java
@@ -1,5 +1,0 @@
-package pl.put.poznan.sqc.logic;
-
-public class Download {
-
-}
