+++ src/main/java/pl/put/poznan/sqc/logic/visitor/ActorAction.java
@@ -1,0 +1,4 @@
+package pl.put.poznan.sqc.logic.visitor;
+
+public class ActorAction {
+}
