+++ src/main/java/pl/put/poznan/sqc/logic/visitor/Download.java
@@ -1,0 +1,5 @@
+package pl.put.poznan.sqc.logic.visitor;
+
+public class Download {
+
+}
