+++ src/main/java/pl/put/poznan/sqc/logic/visitor/MaxDepthScenario.java
@@ -1,0 +1,5 @@
+package pl.put.poznan.sqc.logic.visitor;
+
+public class MaxDepthScenario {
+
+}
